backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv==1.2.1
motor==3.3.1
pydantic==2.12.4
google-genai==1.54.0
httpx==0.28.1
